Look for stage2.img in both locations and fail gracefully. (RBL-4238)

+++ jobslave/generators/installable_iso.py
@@ -267,17 +267,25 @@ class InstallableIso(ImageGenerator):
 
         # extract constants.py from the stage2.img template and override the BETANAG flag
         # this would be better if constants.py could load a secondary constants.py
-        stage2Path = tempfile.mkdtemp(dir=constants.tmpDir)
-        logCall(['/sbin/fsck.cramfs', topdir + '/rPath/base/stage2.img',
-            '-x', stage2Path + '/stage2/'])
-        logCall(['/bin/cp',
-            stage2Path + '/stage2/usr/lib/anaconda/constants.py', tmpPath])
+        for path in ('/images/stage2.img', '/rPath/base/stage2.img'):
+            if os.path.exists(topdir + path):
+                log.info('Copying constants.py from %s', path)
 
-        betaNag = self.getBuildData('betaNag')
-        logCall(['/bin/sed',
-            '-i', 's/BETANAG = 1/BETANAG = %d/' % int(betaNag),
-            tmpPath + '/constants.py'])
-        util.rmtree(stage2Path, ignore_errors = True)
+                stage2Path = tempfile.mkdtemp(dir=constants.tmpDir)
+                logCall(['/sbin/fsck.cramfs', topdir + path,
+                    '-x', stage2Path + '/stage2/'])
+                logCall(['/bin/cp',
+                    stage2Path + '/stage2/usr/lib/anaconda/constants.py',
+                    tmpPath])
+
+                betaNag = self.getBuildData('betaNag')
+                logCall(['/bin/sed',
+                    '-i', 's/BETANAG = 1/BETANAG = %d/' % int(betaNag),
+                    tmpPath + '/constants.py'])
+                util.rmtree(stage2Path, ignore_errors = True)
+                break
+        else:
+            log.info('Could not find stage2.img; not changing beta nag')
 
         # create cramfs
         logCall(['/usr/bin/mkcramfs', tmpPath, productPath])
